+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 Each tagged release MUST have a section 2 heading starting at the time of release `## TAG-NAME...` or Gitlab release notes will be missed.
 
+## v0.21.0
+
+- Development environment (`mw docker`):
+  - Added check to see if docker is running before commands execute (T329920)
+  - Fixed slow DNS lookups when disconnected from the internet, which caused slow MediaWiki requests (T326735)
+  - Added `mediawiki doctor` checks:
+    - Check if `vendor` directory exists (T330926)
+    - Check if a site has been installed (T330928)
+  - Image updates:
+    - docker-registry.wikimedia.org/releng/quibble-buster-php81:1.4.7-s3 to 1.5.1
+    - docker-registry.wikimedia.org/wikimedia/mediawiki-libs-shellbox images to 2023-02-24-002648
+
 ## v0.20.0
 
 - Added XDG standards usage for config directory location (T305150)